어드민에서 쓰인 컴포넌트와 타입을 별도 파일로 이동
